refactor: dissociation of cpcdos api and cpcdos gui api.
