+++ OS2.2/API/includes/CpcdosAPI.h
@@ -13,62 +13,46 @@
 // #include <unistd.h>
 #include <algorithm>
 
-namespace CpcdosAPI {
-    class Window {
-    public:
-        std::string windowName;
-        int windowType;
-        std::tuple<int, int> windowPos;
-        std::tuple<int, int> windowSize;
-        std::tuple<int, int, int, int> windowColor; // RGBA
-        std::tuple<int, int, int> windowTitleColor;
-        std::string windowSettings;
-        std::string windowHandle;
+/*
+==============================================
+=== CPCDOS C++ API
+=== WRITE BY CHRAPATI AND IPLUX
+==============================================
+=== CPCDOSAPI.H
+==============================================
+=== In this file, we set a possibility
+=== to use Cpcdos OSx co-kernel features
+=== in C++ for ehanced developpers.
+==============================================
+=== NO WARANTY TO GET A STABLE PROGRAM DURING
+=== THE FIRST TRY!
+==============================================
+*/
 
-        Window(std::string name = "Window", 
-               int type = 0, 
-               std::tuple<int, int> pos = std::make_tuple(0, 0),
-               std::tuple<int, int> size = std::make_tuple(300, 200), 
-               std::tuple<int, int, int> wincol = std::make_tuple(255, 255, 255),
-               std::tuple<int, int, int> titlecol = std::make_tuple(100, 100, 100),
-               std::string sets = "",
-               std::string handle = "window"
-        );
-    };
-
-    class Textblock {
-    public:
-        std::string textblockText;
-        std::tuple<int, int> textblockPosition;
-        std::tuple<int, int> textblockSize;
-        std::tuple<int, int, int> textColor;
-        std::tuple<int, int, int> backgroundColor;
-        std::string textblockSettings;
-        Window parentWindow;
-        std::string textboxHandle;
-        
-        Textblock(std::string text = "", 
-                  std::tuple<int, int> pos = std::make_tuple(0, 0),
-                  std::tuple<int, int> size = std::make_tuple(100, 100), 
-                  std::tuple<int, int, int> textCol = std::make_tuple(255, 255, 255), 
-                  std::tuple<int, int, int> backCol = std::make_tuple(0, 0, 0), 
-                  std::string sets = "",
-                  Window parWin = nullptr, 
-                  std::string handle = "textblock"
-        );
-    };
-
-    class Picturebox {
-    public:
-        std::string pictureboxHandle;
-        std::string pictureboxFilePath;
-        std::tuple<int, int> pictureboxPosition;
-        std::tuple<int, int> pictureboxSize;
-        int pictureboxDisplayType;
-        Window parentWindow;
-    };
-
+namespace CpcdosAPI 
+{
+    // SHELL FUNCTIONS
     void terminal_print(std::string str);
+
+    // CPCDOS VARIABLES
+
+    // GETTERS
+    void get_cpuid();
+    void get_ram();
+    void get_hour();
+    void get_date();
+    void get_globalTime();
+    void get_processList();
+    void get_threadList();
+
+    // SETTERS
+    void set_activeOS(std::string active_os);
+    
+    // THREADS AND PROCESS
+
+    // CREATING
+    void create_newThread(std::string threadTID, std::string parent_process);
+    void create_newProcess(std::string processPID);
 }
