+++ OS2.2/API/main.cpp
@@ -1,9 +1,11 @@
-#include "includes/CpcdosAPI.h"
-
+// #include "includes/CpcdosAPI.h"
+// 
 // myThing => variable
 // my_thing => function
 // MyThing => namespace
 
 
+
+// This is just an example.
